+++ analysis/src/CompletionDecorators.ml
@@ -175,6 +175,24 @@ Note: The `@react.component` decorator requires the `jsx` config to be set in yo
 
 You will need this decorator whenever you want to use a JSX component in ReScript JSX expressions.|};
       ] );
+    ( "react.componentWithProps",
+      None,
+      [
+        {|The `@react.componentWithProps` decorator is used to annotate functions that are RescriptReact components.
+
+You will need this decorator whenever you want to use a ReScript / React component in ReScript JSX expressions.
+
+Note: The `@react.componentWithProps` decorator requires the `jsx` config to be set in your `rescript.json` to enable the required React transformations.
+
+[Read more and see examples in the documentation](https://rescript-lang.org/syntax-lookup#react-component-with-props-decorator).|};
+      ] );
+    ( "jsx.componentWithProps",
+      None,
+      [
+        {|The `@jsx.componentWithProps` decorator is used to annotate functions that are JSX components used with ReScript's [generic JSX transform](https://rescript-lang.org/docs/manual/latest/jsx#generic-jsx-transform-jsx-beyond-react-experimental).
+
+You will need this decorator whenever you want to use a JSX component in ReScript JSX expressions.|};
+      ] );
     ( "return",
       Some "return(${1:nullable})",
       [