+++ CHANGELOG.md
@@ -35,6 +35,7 @@
 - Dedicated error message for passing a braced ident to something expected to be a record. https://github.com/rescript-lang/rescript/pull/7806
 - Hint about partial application when missing required argument in function call. https://github.com/rescript-lang/rescript/pull/7807
 - More autocomplete improvements involving modules and module types. https://github.com/rescript-lang/rescript/pull/7795
+- Autocomplete `@react.componentWithProps` attribute. https://github.com/rescript-lang/rescript/pull/7812
 
 #### :house: Internal